Revise the process of disaggregation.
The process of disaggregation has been revised to only consider electric home appliances that are present in the household, as indicated in the metadata. Additionally, a normalization step has been introduced, where each home appliance now has a maximum scale value that is used for preprocessing.

For instance, for the predicted value of a fridge appliance, at each timestamp, the predicted value is first set to the maximum of either 0 or the predicted value itself. Then, it is scaled by the ratio of the maximum scale value for the fridge appliance and the maximum predicted value for the entire timeseries (for the entire day). This normalization step ensures that the predicted values are consistent in scale across various home appliances.
pred_fridge = max (0,pred_fridge)
pred_fridge = pred_fridge  *    max_scale_fridge / max_pred_fridge

Furthermore, an additional sensor (noise) is now incorporated into the EnergyData. This sensor represent the difference between the overall consumption (measured by the emonpi) and the sum of the values that have been passed for the disaggregated devices.

Lastly, some absolute imports have been replaced with relative imports.