tcl_asyncdelete async handler fix
Resolved an infrequent bug that occurred when there was a high volume of requests for the Matplotlib plot. The issue was related to a RuntimeError where the tcl_asyncdelete async handler was being deleted by an incorrect thread. Please refer to the following Stack Overflow thread for more information: https://stackoverflow.com/a/29172195.